+++ src/QSymbolicsBase/predefined_fock.jl
@@ -16,20 +16,20 @@ end
 symbollabel(x::FockState) = "$(x.idx)"
 
 """Coherent state in defined Fock basis."""
-@withmetadata struct CoherentState <: AbstractSingleBosonState
-    alpha::Number # TODO parameterize
+@withmetadata struct CoherentState{T<:Number} <: AbstractSingleBosonState
+    alpha::T
 end
 symbollabel(x::CoherentState) = "$(x.alpha)"
 
 """Squeezed vacuum state in defined Fock basis."""
-@withmetadata struct SqueezedState <: AbstractSingleBosonState
-    z::Number
+@withmetadata struct SqueezedState{T<:Number} <: AbstractSingleBosonState
+    z::T
 end
 symbollabel(x::SqueezedState) = "0,$(x.z)"
 
 """Two-mode squeezed vacuum state, or EPR state, in defined Fock basis."""
-@withmetadata struct TwoSqueezedState <: AbstractTwoBosonState
-    z::Number
+@withmetadata struct TwoSqueezedState{T<:Number} <: AbstractTwoBosonState
+    z::T
 end
 symbollabel(x::TwoSqueezedState) = "0,$(x.z)"
 
@@ -112,8 +112,8 @@ julia> qsimplify(phase*c, rewriter=qsimplify_fock)
 |1.2246467991473532e-16 - 1.0im⟩
 ```
 """
-@withmetadata struct PhaseShiftOp <: AbstractSingleBosonGate
-    phase::Number
+@withmetadata struct PhaseShiftOp{T<:Number} <: AbstractSingleBosonGate
+    phase::T
 end
 symbollabel(x::PhaseShiftOp) = "U($(x.phase))"
 
@@ -130,8 +130,8 @@ julia> qsimplify(displace*f, rewriter=qsimplify_fock)
 |im⟩
 ```
 """
-@withmetadata struct DisplaceOp <: AbstractSingleBosonGate
-    alpha::Number
+@withmetadata struct DisplaceOp{T<:Number} <: AbstractSingleBosonGate
+    alpha::T
 end
 symbollabel(x::DisplaceOp) = "D($(x.alpha))"
 
@@ -153,25 +153,25 @@ julia> qsimplify(S*vac, rewriter=qsimplify_fock)
 |0,π⟩
 ```
 """
-@withmetadata struct SqueezeOp <: AbstractSingleBosonGate
-    z::Number
+@withmetadata struct SqueezeOp{T<:Number} <: AbstractSingleBosonGate
+    z::T
 end
 symbollabel(x::SqueezeOp) = "S($(x.z))"
 
 """Thermal bosonic state in defined Fock basis."""
-@withmetadata struct BosonicThermalState <: AbstractSingleBosonOp
-    photons::Number
+@withmetadata struct BosonicThermalState{T<:Number} <: AbstractSingleBosonOp
+    photons::T
 end
 symbollabel(x::BosonicThermalState) = "ρₜₕ($(x.photons))"
 
 """Two-mode squeezing operator in defined Fock basis."""
-@withmetadata struct TwoSqueezeOp <: AbstractTwoBosonGate
-    z::Number
+@withmetadata struct TwoSqueezeOp{T<:Number} <: AbstractTwoBosonGate
+    z::T
 end
 symbollabel(x::TwoSqueezeOp) = "S₂($(x.z))"
 
 """Two-mode beamsplitter operator in defined Fock basis."""
-@withmetadata struct BeamSplitterOp <: AbstractTwoBosonGate
-    transmit::Number
+@withmetadata struct BeamSplitterOp{T<:Number} <: AbstractTwoBosonGate
+    transmit::T
 end
 symbollabel(x::BeamSplitterOp) = "B($(x.transmit))"